requirements.txt
boto3==1.17.64
pem==21.2.0
PyJWT==2.1.0
cryptography==3.4.7
